+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 * The columns defined in ``OAuthConsumerMixin`` now set ``nullable=False``.
   If you are using the SQLAlchemy storage and are upgrading from a previous
   version of Flask-Dance, you may want to do a database migration.
+* Replace ``lazy`` dependency with
+  `werkzeug.utils.cached_property <http://werkzeug.pocoo.org/docs/0.14/utils/#werkzeug.utils.cached_property>`__
 
 `1.4.0`_ (2019-02-22)
 ---------------------